Refs #1: Refactor code to create separate component for Grid
